[shortcutscroll] Ensure last item is never above scrollview bottom
Set a maximum Y position for the scrollview, so the last list item
will never appear above the bottom of the scrollview.

This is to prevent a shortcut to a list divider near the bottom
of the list causing the last list item to be pushed to the top
of the page, leaving lots of blank space.