[spinnerbar] Enable spinnerbar to be overlaid on elements
Rather than have the spinnerbar decorate an element and then
detach that element when it's stopped, specify an element
to be overlaid by the spinner bar. This element isn't removed
when the spinner bar is removed.

This makes the behaviour of the spinner bar consistent with
the spinner, which is located relative to an element, rather
than replacing it.