[spinnerbar] Inline the refresh method
We don't want to expose the refresh() method on the API, as calling
it multiple times may append duplicate spinnerbar elements to the
DOM. So inline it into start, so the DOM manipulation can only
occur if the spinnerbar isn't already running.

When we have a proper refresh() method, we can expose that.