[popupwindow] Use this.element as basis for measurements
Previously the code used this.elem for measurements. However,
this.elem is a copy of this.element made during object creation:
once created, this.elem no longer references the "real" DOM
element.

So, once rendered (no rendering happens in _create), the dimensions
of the DOM element may change; however, the dimensions of
this.elem won't (as it was a copy of this.element).

By referencing this.element instead, we get a correct report
of the element's size, and the popupwindow correctly encloses
the child.