[spinnerbar] Use setInterval instead of recursive function
The spinnerbar used to use a recursive function call (inside
the definition of the called function) to restart the animation
at the end of the loop. This seems to cause a big problem
in the Android 2.2 browser on Galaxy S2.

Instead, use setInterval (with the same duration as the animation)
to repeatedly call the animation in a loop.

The spinnerbar still doesn't work on Android (the CSS used
to draw the barberpole stripes isn't supported), but at least
the page doesn't crash altogether.