header: Move button top higher in header with tabbar
Buttons in the page header have its own position(which is described in
the Tizen GUI guide), but if the button and the tabbar are put into the
header together, the tabbar overlaps the lower part of buttons. Although
there is no specification about the vertical position of the buttons
when the tabbar is used together, this patch moves the button slightly
higher, not to be overlaped by the tabbar.
Resolves N_SE-22253.

Change-Id: I92c61eaf10ca6134d1b625da5bd3d1e71aa4c84d