pagelayout: Remember back key when it is hidden
If an app moves a page when the pop is shown, the .ui-page-active page is changed.
In this case, the '.ui-page-active .ui-btn-back' selector indicates a new page's
one, and the old page's .ui-btn-back button is still hidden.
So, this patch remember the current back button, and show it again at the
softkeyboardchange.off event.

Resolves #N_SE-32900.

Change-Id: I94fd641ca0e70b4093b881e2319c6f18a2b912a6