JIRA - FIX N_SE-34055 Resolving problem with the pinch event trigered on the image after orientation change
Resolving problem with the pinch event triggered on the image tag after orientation change.

The problem is with the image placed in the gallery. If the orientation is changed then the gallery remove image and insert the new one without properties assigned by pinch application.

Pinch event has scale down minimum set to 0.1. This limit is remembered after rotation with the full size image(the one replaced by gallery), disabling the pinch event.

Change-Id: I1f54f940de6c3775f85df113d0dd6eb9bee0f9d1