pagelayout : fixed triggering blur event after pressing H/W menu key (connected with N_SE-47012)
When blur event is triggered, the .ui-focus class is removed from focused element, so checking
if focused element is in popup window after triggering blur event doesn't have sense.

This checking should be before triggering blur event and (according to comment in lines 333-335)
if focused element is in popup window, the popup won't be closed then.

Change-Id: I2ce68316287897f0ce32a982d2951de12110f573