JQM path: modify vclick move distance threshold and calulation way.(P131106-04132,N_SE-56430)
The JQM vmouse events provides vclick, and it uses touchstart and touchend.
To create this virtual event, it catches event.pageX and event.pageY by
default. These value are based on the scaled webview width and height,
determined by the viewport meta tag.
The vclick event allows the threshold distance between touchstart and
touchend.  Similary, the click event fired by the browser also allows
same kind of threshold.  If the distance is longer than the threshold,
the event is not a vclick or click.
However, the threshold is differeent between vclick and click.
So, in this commit, the event.pageX and event.pageY attributes used by
vclick calculation is replaced by the event.screenX and event.screenY,
to have same metric to webkit click event. Also, the distance
calculation algorithm is changed by using Pythagoras Theorem.

Change-Id: I0e840d18cd0df154431342fa06df2b764d689460