[CORE] Popup: change position of arrow
If we put id of element or HTML element as a value of property positionTo in popup,
the coordinates x and y will be undefined. In such situation, the arrow will be shown,
but the position of arrow will not be set.

In this fix we try to emulate the position of event by counting the proper values of x, y.
These values depend on the type of arrow, because we get the center of the proper edge
of element from property positionTo.

Change-Id: I98ba4d26b7028b5d919d7a3b81188444b6d83cbb
Signed-off-by: Jadwiga Sosnowska <j.sosnowska@samsung.com>