feat(mobile): KIRAN SelectMenu widget
This is SelectMenu widget for new theme.
The instructions are almost like JQM.
Some options are removed.
1) theme, corners, shadow, iconshadow, overlayTheme, closeText,
mini options are removed since we will not support.
2) icon, iconPos, elementsToDialog options are also removed
since the UX requires a specific icon and rule.
3) border-radius of dropdown list is the UX requirement
but it is not applied owing to problem of webkit rendering
for now.

API-Change: mobile/selectmenu
Change-Id: Iff547d39f4c5c1f2ee5207f3436f59f66b2c64e2
Signed-off-by: hagun.kim <hagun.kim@samsung.com>