[Mobile Multi Page Sample] fix receiving back key event when popup is opened.
in mobile profile, tau catch back key event
and if popup is opend, it close popup and stop propagation key event.
but this main routine catches key event same level(document element) with tau.
so, I lowered the level of the element so that this main routine can't receive the key event.

Change-Id: I1858eac1c5efa81f73c4e859317f53e6df5b5a89
Signed-off-by: Hyunkook Cho <hk0713.cho@samsung.com>