fix(core): fix problem with opening Popup by function in Router
There was no way to open popup without closing the previous one.
When one popup was opened and we called function tau.openPopup(), all popups
(the previous and the new one) were closed, because of the hash change.

This problem was fixed in this commit and now it is possible to open
a new popup without closing the previous one manually.

Change-Id: I2415f1d0fc71910834d61046c672e7a3d9e6cad2
Signed-off-by: Jadwiga Sosnowska <j.sosnowska@samsung.com>