feat(core): add module with decorator for animation on events focus/blur
The new namespace "decorator" was added in core and the decorator
for animation on events focus/blur was prepared.

Now, if it is possible to focus/blur widget's element, the event
"taufocus"/"taublur" is triggered and the decorator focusAnimation
adds/removes proper css classes.

It is used only in TV profile.

Change-Id: Ib80ed4bc9e2e3321d2992fdea9d86fa065e7d5b1
Signed-off-by: Jadwiga Sosnowska <j.sosnowska@samsung.com>