etc(tool): set up build-bot with jenkins server
We set up the build-bot for web-ui-fw git project.
It helps to automate building, testing and etc.
When new patchset is created on gerrit, below tasks are executed.

[build]
- grunt build

[unit test]
- grunt test

[jshint]
- grunt jshint:single --jshintfile=<file>
- jshint uses source files that are located in "tau/src/js/".
- jshintrc options will be discussed and fixed, before then jshint result doesn't affect gerrit review.

[code coverage]
- The TCs of each profile reference same path of tau.js (tau/tests/libs/dist/js/tau.js). So code coverage can not be tested for each profile.
- If this issue is fixed, code coverage is tested. But coverage result doesn't affect gerrit review.

When all tasks are executed, the system will push comment and verified value on gerrit.
Verified value depends on build and test result.

Change-Id: Ic733488518dc2212a5951e2296c7cce5b7f334b8
Signed-off-by: Junyoung Park <jy-.park@samsung.com>