fear(wearable): page scroll on default
By this commit,
the usage of "ui-scroll-on" is converted to a default.

And the default state of the usage is set to "false", which means
the class "ui-scroll-on" is *not* in use.

To change the usage backwards(as positive) for the whole page,
please change default value.

To change the usage backwards(as positive) for a specific page,
please use 'data-enable-page-scroll = "true"' into the page.

Change-Id: I77ac9cb35379a78acf7baa8c045a6772cfd94373
Signed-off-by: junhyeonLee <juneh.lee@samsung.com>