refactor(mobile): integrate FastScroll and IndexScrollbar
The FastScroll widget is not fully replaceable by the IndexScrollbar.
The IndexScrollbar is stand alone widget with collection of shortcuts
and the widget triggeres "select" event when user select a char.

A developer has to bind the IndexScrollbar with other widgets while
the FastScroll extends the Listview then is automatically bound with
the Listview.

Change-Id: Icf5d8cb8fb093f1411a50c63e2c741ed8bd4a0dc
Signed-off-by: Tomasz Lukawski <t.lukawski@samsung.com>
Signed-off-by: Piotr Karny <p.karny@samsung.com>