fix(tv): changing context after opening popup
There was problem with focus in case of popups without links,
which were not toast. The context was not changed and the focus
stayed on page after opening popup.

Now, the context is changed for all kinds of popups apart from
popups connected with slider. Moreover such popups, which don't have
focusable elements inside, can be closed by event keyup.

Change-Id: I00f08f8a3494e8a67c3864e80f61a9bfb73ae985
Signed-off-by: Jadwiga Sosnowska <j.sosnowska@samsung.com>