feat(mobile): add structure for layouts
There was problem with adding new layout, because layouts
didn't have any structure. All kinds of layouts were implemented
in the same file, where widget Box was defined.

In this commit the stucture for layouts was proposed. Now,
a new layout can be defined in a separated file and added to
special namespace (tau.layout). To be ready for use, it has
to register itself in widget Box.

Change-Id: I010ff64c37c55225193fdc670cec7850cc749cae
Signed-off-by: Jadwiga Sosnowska <j.sosnowska@samsung.com>