feat(mobile): add toggleswitch style for 2.4
The toggle style for 2.4 mobile is added.

Previously, toggle is based on label and div,
and input was used to check state, not to show.

Now it is changed to container and input directly,
so input is now visible as toggle.
-------------------------------------------------
From patch 5,

ToggleSwitch with <select data-role="slider"> is
separated from usual toggleswitch.
And the slider version is save as :
ToggleSwitch.extra.js
toggleswitch.extra.less

So they are easy to be removed later.

Change-Id: Ibc90f9a7cb001dca6dc8dda5a743096844dbe610
Signed-off-by: junhyeonLee <juneh.lee@samsung.com>
Signed-off-by: Gahye Kim <gahye07.kim@samsung.com>