tests(mobile) Add textarea inside ui-test Popup to pass
---
Problem:
Need of a way to verify fix of textarea height inside Popup.

---
Solution:
Earlier bug with textarea height inside Popup was fixed and ui-test
is added to pass to ensure that in future it will not be a problem.
Parameter "pass" has been changed to true in configuration file.

---
Possible SideEffects:
Should be none.

---
Verification:
Run ui-test.

 - OAPTAU-1049

Change-Id: Ic9c59b41c696464a8dc4fc927b04c99d630b58fb
Signed-off-by: p.kepinski <p.kepinski@samsung.com>
Reviewed-on: http://AMDC2476:8080/140
Reviewed-by: Maciej Urbanski <m.urbanski@samsung.com>
Tested-by: Maciej Urbanski <m.urbanski@samsung.com>