fix(wearable) Make scrollbar visible on VirtualListSimple
Problem:
Scrollbar in circle version was not visible.
On rectangle where Scrollview is not available there was an exception on
list update.

Solution:
Structure of scrollbar was correctly built in SVG but had wrong colors.
Added condition to eliminate exception.
Start render loop on rotary.

Verification:
Samples for virtual list on wearable in circle and rectangle mode.

Change-Id: I5c407e81ea5a35c055fe623c329c7303513152ef
Reviewed-on: http://AMDC2476:8080/160
Reviewed-by: Dariusz Dyrda <d.dyrda@samsung.com>
Tested-by: Jenkins
Reviewed-by: Maciej Urbanski <m.urbanski@samsung.com>