tests(mobile) Increase test coverage
Problem:
New lines in Popup.js are not covered. Build method is not tested.

Solution:
Write new tests. Adding class method changed in mobile Popup module to
make the tests pass. apply() function with array passed as parameter
made tests fail so it was changed to for loop with the same
functionality.

Verification:
Run tests.

Other:
 - OAPTAU-1069

Change-Id: I9182483e3f9b2665716a626a02e6aec3b31fda20
Signed-off-by: p.kepinski <p.kepinski@samsung.com>
Reviewed-on: http://AMDC2476:8080/185
Tested-by: Jenkins
Reviewed-by: Maciej Urbanski <m.urbanski@samsung.com>
Tested-by: Maciej Urbanski <m.urbanski@samsung.com>