tests(core) Fix unit test for _initStaticAndWidths method in Tabbar
Problem:
This particular test was failing because it was dependent on screen
width of the client the test was running on.

Solution:
Added inline styles that limit width of the tabs.

Verification:
Run unit tests in Karma. It may be necessary to comment out all other
tests except for Tabbar.

Change-Id: I1a05fbd4fa29e88ec81a810312b35120a25c6989
Signed-off-by: Dariusz Dyrda <d.dyrda@samsung.com>
Reviewed-on: http://AMDC2476:8080/195
Tested-by: Jenkins
Reviewed-by: Maciej Urbanski <m.urbanski@samsung.com>